+++ source/archetypes/_archetypeDictionary.js
@@ -7,6 +7,7 @@ for (const file of [
 	"assassin.js",
 	"chemist.js",
 	"detective.js",
+	"fighter.js",
 	"hemomancer.js",
 	"knight.js",
 	"legionnaire.js",